+++ src/interfaces/UsdnProtocol/IUsdnProtocol.sol
@@ -1,6 +1,8 @@
 // SPDX-License-Identifier: MIT
 pragma solidity >=0.8.0;
 
+import { IUsdnProtocolErrors } from "./IUsdnProtocolErrors.sol";
+import { IUsdnProtocolEvents } from "./IUsdnProtocolEvents.sol";
 import { IUsdnProtocolFallback } from "./IUsdnProtocolFallback.sol";
 import { IUsdnProtocolImpl } from "./IUsdnProtocolImpl.sol";
 
@@ -8,7 +10,7 @@ import { IUsdnProtocolImpl } from "./IUsdnProtocolImpl.sol";
  * @title IUsdnProtocol
  * @notice Interface for the USDN protocol and fallback.
  */
-interface IUsdnProtocol is IUsdnProtocolImpl, IUsdnProtocolFallback {
+interface IUsdnProtocol is IUsdnProtocolImpl, IUsdnProtocolFallback, IUsdnProtocolErrors, IUsdnProtocolEvents {
     /**
      * @notice Upgrades the protocol to a new implementation (check
      * [UUPSUpgradeable](https://docs.openzeppelin.com/contracts/5.x/api/proxy#UUPSUpgradeable)).